Helper function to generate cell IDs to be used for choosing the environmental background data

Returns pixel IDs to be sample for generating environmental background form modeling layers.

Usage

cells2samp(
  data,
  longitude,
  latitude,
  cell_ids = NULL,
  buffer_ngbs = 2,
  raster_mask,
  process_ngbs_by = 10,
  n_bg = 50000,
  progress = TRUE
)

Arguments

data

A data.frame with longitude and latitude data

longitude

A character vector of the column name of longitude.

latitude

A character vector of the column name of latitude.

cell_ids

A numeric vector with pixel ids. The default values NULL. Use this parameter if you have obtained this information using the function cellFromXY.

buffer_ngbs

Number of pixel neighbors around occurrences to be used to build the buffer.

raster_mask

An object of class SpatRaster that will be used to obtain pixel IDs.

process_ngbs_by

Numeric. Estimates neighbor cells each x cells. This is for memory management.

n_bg

Number of background pixels.

progress

Logical. Show computation progress

Value

A numeric vector with the IDs of cells to be sampled.

Examples

# \donttest{
# cells to sample
data(abronia)
temporal_layer <- system.file("extdata/bio/2016/bio_01.tif",package = "tenm")
raster_mask <- terra::rast(temporal_layer)
set.seed(123)
samp_01 <- tenm::cells2samp(data = abronia,
                            longitude = "decimalLongitude",
                            latitude = "decimalLatitude",
                            cell_ids = NULL,
                            buffer_ngbs = 4,
                            raster_mask = raster_mask,
                            process_ngbs_by = 10,
                            n_bg = 50000,
                            progress =TRUE)
# }