Function to clean duplicated coordinates data

Clean up duplicated longitude and latitude data by year using a threshold distance. The main propose of this function is eliminate data occurrence points spliced in each different year or date by a given distance threshold. This function also allows you to clean duplicate data by pixel at a given resolution of a raster mask.

Usage

clean_dup_by_date(
  this_species,
  threshold,
  by_mask = FALSE,
  raster_mask = NULL,
  n_ngbs = 0
)

Arguments

this_species

Species Temporal Data object see sp_temporal_data.

threshold

A numeric value representing the distance between coordinates to be considered as a duplicate.

by_mask

Logical. If TRUE the elimination of duplicates will be done using a raster layer as a mask; If False the elimination of duplicates will be done by the distance threshold.

raster_mask

An object of class SpatRaster that will be used to clean duplicates that are present in the same ID pixel.

n_ngbs

Number of pixel neighbors. Remove duplicates depending on how many pixels range you want,"0" is use in order to remove all the duplicated occurrence points present in a single pixel (1x1), "1" is to eliminate duplicates in a 1 pixel-long neighborhood of a 9 pixel area (3x3), "2" correspond to a neighborhood of 2 adjacent pixels of a 25 pixel area (5x5) and so on depending on how much area you want to cover.

Details

This function is build on the basis of clean_dup. See the help of the function for more examples

Examples

library(tenm)
data("abronia")
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
tenm_mask <- terra::rast(file.path(tempora_layers_dir,"1939/bio_01.tif"))
# Clean duplicates without raster mask (just by distance threshold)
abt <- tenm::sp_temporal_data(occs = abronia,
                              longitude = "decimalLongitude",
                              latitude = "decimalLatitude",
                              sp_date_var = "year",
                              occ_date_format="y",
                              layers_date_format= "y",
                              layers_by_date_dir = tempora_layers_dir,
                              layers_ext="*.tif$")
abtc1 <- tenm::clean_dup_by_date(abt,threshold = terra::res(tenm_mask)[1])
# Check number of records
print(nrow(abtc1$temporal_df))
#> [1] 40
# Clean duplicates using a raster mask
abtc2 <- tenm::clean_dup_by_date(this_species = abt,
                                by_mask = TRUE,
                                threshold = terra::res(tenm_mask)[1],
                                raster_mask = tenm_mask[1],
                                n_ngbs = 0)
# Check number of records
print(nrow(abtc2$temporal_df))
#> [1] 40