Function to compute the Minimum Volume covariance Matrix of an ellipsoid niche model.

Function to compute the covariance matrix, the niche centroid and volume of an ellipsoid model. It uses the values of the niche variables of the occurrences points.

Usage

cov_center(data, mve = TRUE, level, vars = NULL)

Arguments

data

A data.frame or a matrix with the numeric values of the variables that will be used to model the niche.

mve

A logical value. If TRUE a minimum volume ellipsoid will be computed using the function cov.mve of the MASS package. If False the covariance matrix of the input data will be used.

level

A numerical value specifying the proportion of the data to be used to compute the ellipsoid.

vars

A numeric or a string vector specifying the columns indexes/names of the variables of the input data which will be used to fit the ellipsoid model. If NULL the user will be asked to enter the indexes. interactively

Value

Returns a list containing the centroid of the ellipsoid, the covariance matrix based on the input data, ellipsoid volume, semi-axis length and axis coordinates.

Examples

# \donttest{
library(tenm)
data("abronia")
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
abt <- tenm::sp_temporal_data(occs = abronia,
                              longitude = "decimalLongitude",
                              latitude = "decimalLatitude",
                              sp_date_var = "year",
                              occ_date_format="y",
                              layers_date_format= "y",
                              layers_by_date_dir = tempora_layers_dir,
                              layers_ext="*.tif$")
abtc <- tenm::clean_dup_by_date(abt,threshold = 10/60)
future::plan("multisession",workers=2)
abex <- tenm::ex_by_date(abtc,train_prop=0.7)
future::plan("sequential")
varcorrs <- tenm::correlation_finder(environmental_data = abex$env_data[,-ncol(abex$env_data)],
                                     method = "spearman",
                                     threshold = 0.8,
                                     verbose = FALSE)
#> Warning: the standard deviation is zero
mod <- tenm::cov_center(data = abex$env_data,
                        mve = TRUE,
                        level = 0.975,
                        vars = c("bio_05","bio_06","bio_12"))
# Print model parameters
print(mod)
#> $centroid
#>     bio_05     bio_06     bio_12 
#>  217.10256   54.69231 1106.07692 
#> 
#> $covariance
#>           bio_05   bio_06      bio_12
#> bio_05  731.3050 456.5061   -249.8502
#> bio_06  456.5061 418.6923    987.1559
#> bio_12 -249.8502 987.1559 152235.5992
#> 
#> $niche_volume
#> [1] 14127753
#> 
#> $SemiAxis_length
#>          a          b          c 
#>   28.44591   99.38665 1192.98935 
#> 
#> $axis_coordinates
#> $axis_coordinates[[1]]
#>         bio_05   bio_06   bio_12
#> vec_1 233.5857 31.50945 1106.254
#> vec_2 200.6194 77.87517 1105.899
#> 
#> $axis_coordinates[[2]]
#>         bio_05     bio_06   bio_12
#> vec_1 298.1029 112.282107 1105.835
#> vec_2 136.1022  -2.897492 1106.319
#> 
#> $axis_coordinates[[3]]
#>         bio_05   bio_06     bio_12
#> vec_1 215.1587 62.44309 2299.03951
#> vec_2 219.0465 46.94153  -86.88566
#> 
#> 
# }