inEllipsoid: Determine if a point is inside or outside an ellipsoid

Determine if a point is inside or outside an ellipsoid.

Usage

inEllipsoid(centroid, eShape, env_data, level)

Arguments

centroid

A numeric vector of centroids for each environmental variable

eShape

Shape matrix of the ellipsoid (can be a covariance matrix or a minimum volume ellipsoid).

env_data

A data frame with the environmental training data.

level

Proportion of points to be included in the ellipsoids. This parameter is equivalent to the error (E) proposed by Peterson et al. (2008).

Value

A data.frame with 2 columns. The first "in_Ellipsoid" binary response with values 1 (inside the ellipsoid) and zeros (outside the ellipsoid); the second "mh_dist" Mahalanobis distance to centroid.

Examples

# \donttest{
library(tenm)
data("abronia")
tempora_layers_dir <- system.file("extdata/bio",package = "tenm")
abt <- tenm::sp_temporal_data(occs = abronia,
                              longitude = "decimalLongitude",
                              latitude = "decimalLatitude",
                              sp_date_var = "year",
                              occ_date_format="y",
                              layers_date_format= "y",
                              layers_by_date_dir = tempora_layers_dir,
                              layers_ext="*.tif$")
abtc <- tenm::clean_dup_by_date(abt,threshold = 10/60)
future::plan("multisession",workers=10)
abex <- tenm::ex_by_date(abtc,train_prop=0.7)
varcorrs <- tenm::correlation_finder(environmental_data = abex$env_data[,-ncol(abex$env_data)],
                                     method = "spearman",
                                     threshold = 0.8,
                                     verbose = FALSE)
#> Warning: the standard deviation is zero
future::plan("sequential")
mod <- tenm::cov_center(data = abex$env_data,
                        mve = TRUE,
                        level = 0.975,
                        vars = c("bio_05","bio_06","bio_12"))
in_elip <- tenm::inEllipsoid(centroid = mod$centroid,
                       eShape = mod$covariance,
                       env_data = abex$env_data[,c("bio_05","bio_06","bio_12")],
                       level = 0.975)
# 1 = Inside the ellipsoid; 0 = Outside the ellipsoid
print(in_elip)
#>    in_Ellipsoid    mh_dist
#> 1             1  1.4705997
#> 2             1  1.1194538
#> 3             1  5.0059315
#> 4             1  0.6661832
#> 5             1  0.6980075
#> 6             1  2.9156649
#> 7             1  2.8671845
#> 8             1  0.5860615
#> 9             1  1.4006391
#> 10            1  0.5028708
#> 11            1  0.9491132
#> 12            1  5.9272317
#> 13            1  0.6541567
#> 14            1  2.6362835
#> 15            1  4.9646196
#> 16            1  0.4645683
#> 17            1  0.9290738
#> 18            1  1.5385982
#> 19            1  0.1544606
#> 20            1  5.1650561
#> 21            1  1.8380630
#> 22            0 10.0864966
#> 23            1  1.7757906
#> 24            1  3.2762505
#> 25            1  6.4194140
#> 26            1  2.0056851
#> 27            1  3.3470657
#> 28            1  2.6451526
#> 29            1  0.8124142
#> 30            1  8.2305333
#> 31            1  2.0567643
#> 32            1  1.4496595
#> 33            0  9.7469290
#> 34            1  0.8046435
#> 35            1  4.8280389
#> 36            1  7.5071430
#> 37            1  4.7907513
#> 38            1  7.0150910
#> 39            1  3.0192994
#> 40            1  1.8155527
# }