Partial ROC calculation for Niche Models

pROC applies partial ROC tests to continuous niche models.

Usage

pROC(
  continuous_mod,
  test_data,
  n_iter = 1000,
  E_percent = 5,
  boost_percent = 50,
  rseed = FALSE,
  sub_sample = TRUE,
  sub_sample_size = 1000
)

Arguments

continuous_mod

a SpatRaster or a numeric vector of the ecological niche model to be evaluated. If a numeric vector is provided it should contain the values of the predicted suitability.

test_data

A numerical matrix, data.frame, or a numeric vector. If it is data.frame or matrix it should contain coordinates of the occurrences used to test the ecological niche model to be evaluated; columns must be: longitude and latitude. If numeric vector it should contain the values of the predicted suitability.

n_iter

(numeric) number of bootstrap iterations to be performed; default = 1000.

E_percent

(numeric) value from 0 to 100 that will be used as a threshold (E); default = 5.

boost_percent

(numeric) value from 0 to 100 representing the percent of testing data to be used for performing the bootstrap process for calculating the partial ROC; default = 50.

rseed

Logical. Whether or not to set a random seed. Default FALSE.

sub_sample

Logical. Indicates whether the test should run using a subsample of size sub_sample_size. It is recommended for big rasters.

sub_sample_size

Numeric. Size of the sample to be used for computing pROC values.

Value

A data.frame containing the AUC values and AUC ratios calculated for each iteration.

Details

Partial ROC is calculated following Peterson et al. (2008; http://dx.doi.org/10.1016/j.ecolmodel.2007.11.008). This function is a modification of the PartialROC funcion, available at https://github.com/narayanibarve/ENMGadgets.

References

Peterson, A.T. et al. (2008) Rethinking receiver operating characteristic analysis applications in ecological niche modeling. Ecol. Modell., 213, 63–72.

Examples

data(abronia)
suit_1970_2000 <- terra::rast(system.file("extdata/suit_1970_2000.tif",
                                          package = "tenm"))
print(suit_1970_2000)
#> class       : SpatRaster 
#> dimensions  : 23, 20, 1  (nrow, ncol, nlyr)
#> resolution  : 0.1666667, 0.1666667  (x, y)
#> extent      : -99.16667, -95.83333, 17.16667, 21  (xmin, xmax, ymin, ymax)
#> coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#> source      : suit_1970_2000.tif 
#> name        : suit_1970_2000 
#> min value   :   4.614139e-37 
#> max value   :   8.332142e-01 
proc_test <- tenm::pROC(continuous_mod = suit_1970_2000,
                        test_data = abronia[,c("decimalLongitude",
                                               "decimalLatitude")],
                        n_iter = 500, E_percent=5,
                        boost_percent=50)
print(proc_test$pROC_summary)
#>              Mean_AUC Mean_pAUC_ratio_at_5%               P_value 
#>             0.9024062             1.5066030             0.0000000